requirements.txt
gunicorn
dj-database-url
whitenoise
django-cors-headers